requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
python-dotenv==1.0.0